backend/requirements.txt
fastapi
uvicorn
sqlalchemy
openai
python-dotenv
requests
pydantic
Pyyaml